impl(gax): prevent `ClientConfig` breaking changes (#3284)
This type is in the internal namespace, and `#[doc(hidden)]` so it is
not part of the public API for two reasons. However, nothing prevented
us from using the type in a way that can introduce breaking changes.

Making the type `#[non_exhaustive]` forces us to use the type in more
backwards compatible ways.