feat(storage): add rewrite_until_done helper as extension trait (#3328)
Several examples for object rewrite operations required a manual loop to poll the operation until it was complete. This boilerplate is cumbersome for users and prone to error.

This change introduces a `RewriteObjectExt` extension trait, which provides `rewrite_until_done()` helper method on the `RewriteObject` builder. This function encapsulates the polling logic, automatically handling the rewrite_token and looping until the operation is finished.

This extension trait lives in the new `builder_ext` module. This is modeled after the `model_ext` module which contains other convenience functions and types that are built on top of generated types.

For #3204