test(pubub): add cleanup for topics (#3363)
Cleanup stale topics created by integration tests. Pubsub topics don't have a create time field, so instead create a label that records the creation time. There are restrictions on the characters used in labels including no uppercase letters. The wkt::Timestamp serialization uses uppercase letters, so instead this keeps it simple for the tests and stores the time in seconds.