fix: support locational endpoints (#3402)
Part of the work for #3375. (The other half is gRPC).

When a client is configured to use a locational / regional endpoints,
use them as the host, instead of the default host (global). The client
library detects these automatically.

Adds an integration test that fails without the corresponding changes to
the `gaxi` client.