feat(gax): Add public extensions to Response::Parts and propagate from transports
This change introduces a public extensions field of type http::Extensions
to the gax::response::Parts struct.

The transport implementations in gax-internal for both HTTP (reqwest)
and gRPC (tonic) have been updated to clone the extensions from the
underlying transport response into this new field.

This allows metadata and other information to be passed through the
response chain via the extensions mechanism.

Tests have been added to verify the propagation in both gRPC and HTTP transports.