docs: update generate instructions to use GOPROXY=direct
We want to be able to update the generated code immediately after
submitting code to github. The proxy will cache results for lists
and branch resolutions. To bypass this, use GOPROXY=direct to bypass
the proxy when getting the version. We can then use that version
directly with proxy, which should then go fetch that version from
github.