docs: update generate instructions to use GOPROXY=direct (#3492)
We want to be able to update the generated code immediately after
submitting code to Github. The proxy will cache results for lists and
branch resolutions. To bypass this, use GOPROXY=direct to bypass the
proxy when getting the version and for `go run`.

The downloaded version of sidekick will still be verified against the
checksum database.

It is possible that the versions will differ if there is a commit in
between running the two commands, but our CI will catch that.