refactor(auth): service account access token generation (#3523)
Create reusable `ServiceAccountTokenGenerator` to be used by both access
token provider and id token provider. Also add `target_audience` to
`JwsClaims`, which is needed for ID Tokens. The generator is used for
two purposes:
* For Access Token provider passes `scopes` and `audience` configurable
by user. The generated access token is returned to consumer. The
`target_audience` field is not used.
* For ID Token provider pass `audience =
"https://oauth2.googleapis.com/token"` and `target_audience` is the end
audience for the id token. The generated access token then is used to
request a ID Token on `https://oauth2.googleapis.com/token`. The
`scopes` field is not used.

Towards #3449