refactor(gax-internal): Use functions for HTTP spans (#3637)
Removing HttpSpanInfo simplifies the implementation and removes a step
of packing and unpacking data.

The downside is that types are now implicit: we could accidentally log
an i16 as a String and TestLayer / tests would not be aware of the
problem. I plan to add this to TestLayer in a future PR to improve the
tests.