impl(storage): handle redirect errors (#3715)
In the bidi streaming protocol the stream may be closed to redirect the
client to a new handle and routing header. This PR introduces a helper
function to convert a `grpc::Status` to `gax::error::Error` **and** also
update the read spec (part of the initial bidi read request) with any
redirect data.

Parsing the error requires several storage-specific types. It seems
better to keep this in the storage library.