feat(pubsub): add `flush` to the publisher
This adds a flush command to the Publisher, providing a reliable
way for users to force all messages to send. This is accomplished
by sending the flush request through the same channel as messages
are requested, ensuring that any publish request received before
flush is included in the flush.

For #3015