test(pubsub): add a test for dropping the publisher (#3726)
The added test verifies that the messages are sent with no delay when
the publisher is dropped.

Dropping the publisher should trigger any remaining messages to send. If
the user is still holding on to the handles, this will prevent the
program from ending and killing the spawned task that handles the send.