impl(storage): `open_object()` request builder
This is not very interesting code, just wrappers around other types. I
think it is easier to review and will reduce the size of future PRs.

Notably, there are no setters for `routing_token` or `read_handle`, as
this will be part of the public API and should not expose those
implementation details. Notably, it uses a tonic-generated type
directly, this may need to change when we introduce a mocking layer.