feat(pubsub): add `flush` to the publisher (#3728)
This adds a flush command to the Publisher, providing a way for users to
force all messages to send without waiting for batching limits. This is
implemented by sending the flush request through the same channel as
messages are requested, ensuring that any publish request received
before flush are included in the flush.

For #3015