impl(storage): handle pending ranges
A `PendingRange` will be used in the implementation of bidi streaming reads. As
the name implies, it represents a single range read that still has some data
pending. It verifies the data arrives in order, and that the checksums pass. If
there are any problems the caller should terminate all ranged reads in that
stream, as some catastrophic error has ocurred.