impl(storage): a `ReadObjectResponse` for ranges
In a bidi streaming read the application will call `open_object()`, get back an
object descriptor, and then call `read()` on the descriptor to read a range.
The return from this `read()` is the same type as calling `read_object()` on
the client. The implementation is a different type, implemented here.