tests: Mark some examples as compile_fail (#3911)
We will mark generated examples as ignore because they take a long time
to build. Still we want to test them weekly using `--ignored` which
builds and runs all ignored examples.

The examples on this PR were previously marked as ignore because they
don't build by design. If we keep them as are they will fail when we use
`--ignored` to run ignore samples, so we mark them as compile_fail
instead.

Towards #2808