fix(gax-internal): no TLS config for http and IPv6
When running storage tests the `httptest` module may pick an IPv6 address. The
storage client tries to create an (unused) grpc::Client for the test endpoint,
and fails because tonic rejects `http://[::]:*` addresses with TLS.

I suspect a bug in tonic, because IPv4 addresses seem to work. Regardless, we
need this case to work to keep the complexity of our tests low.