impl(pubsub): nascent lease management (#3971)
Part of the work for #3957 

Add a crate private struct to help track the lease state.

A follow up PR will introduce a `trait Leaser`, and start taking actual
actions (on `flush()`).

The plan is to keep at-least-once and exactly-once delivery mostly
separate. The code in this PR only applies to at-least-once delivery.