test(pubsub): mock gRPC server for testing (#4083)
Part of the work for #3957 

Follow the lead of #3932 (and basically copy all of the code but adapt
it for Pub/Sub). We will use it to cover the handwritten concrete
implementation of `streaming_pull`.

Note that:
- integration tests against production will not do. I need to test a
`pub(crate)` type.
- tests against the emulator are hard to set up.
- we don't want to block pubsub progress on a generic solution for Bidi
across all the GAPICs

This code is fairly mechanical to write, but it is too early to think
about generating it with sidekick.