impl(pubsub): lease loop (#4120)
Part of the work for #3957 

Add a lease loop. The loop runs on end, performing any of the following
actions:

- handle a shutdown
- flush acks/nacks
- extend leases
- accept a new message
- handle incoming acks/nacks

I combined these things into a struct, but I may change my mind later.

## On testing...

TIL: `mock.checkpoint()`. It means check all the expectations at this
exact moment.

I found it easiest to use an `Arc<Mutex<Mock>>` which allowed me to give
the mock to the lease loop AND set expectations on the mock outside of
the lease loop.

I also tried to recover the time at which events happen from within the
mock expectations, but decided that was more complicated.

Because `tokio::test` is single threaded (if you want to `start_paused =
true`), I found that I needed to yield the current execution so tokio
would actually launch the lease loop. Otherwise it would not launch the
lease loop until the first `.await` point, and the timing was all messed
up.